+++ core/common/src/main/java/com/emotionstorage/common/util/SerializerUtil.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.common
+package com.emotionstorage.common.util
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
