+++ core/common/src/main/java/com/emotionstorage/common/util/YearMonthUtil.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.common
+package com.emotionstorage.common.util
 
 import java.time.LocalDate
 import java.time.YearMonth
