+++ core/remote/src/main/java/com/emotionstorage/remote/response/dailyReport/GetDailyReportResponse.kt
@@ -1,6 +1,6 @@
 package com.emotionstorage.remote.response.dailyReport
 
-import com.emotionstorage.common.LocalDateTimeSerializer
+import com.emotionstorage.common.util.LocalDateTimeSerializer
 import kotlinx.serialization.Serializable
 import java.time.LocalDateTime
